requirements.txt
flask>=3.0.0
Werkzeug>=3.0.0
google-generativeai>=0.3.2
Pillow>=10.0.0
python-dotenv>=1.0.0
pdf2image>=1.16.3
markdown>=3.5.2 